Move stubApiRequests to lib
